Completed Project clone and removing references from HERE Studio file
